+++ lib/src/clojure_lsp/http.clj
@@ -6,16 +6,16 @@
 
 (set! *warn-on-reflection* true)
 
-(def ^:private connect-timeout-ms 1000)
-(def ^:private read-timeout-ms 30000)
+(def ^:private default-connect-timeout-ms 1000)
+(def ^:private default-read-timeout-ms 30000)
 
 (defn test-remote-url! [^String url]
   (if-not (.startsWith url "http")
     [true]
     (let [url (URL. url)
           conn ^HttpsURLConnection (.openConnection url)]
-      (.setConnectTimeout conn connect-timeout-ms)
-      (.setReadTimeout conn read-timeout-ms)
+      (.setConnectTimeout conn default-connect-timeout-ms)
+      (.setReadTimeout conn default-read-timeout-ms)
       (try
         (.connect conn)
         [true]
@@ -28,8 +28,8 @@
   (try
     (let [url (URL. url)
           conn ^HttpsURLConnection (.openConnection url)]
-      (.setConnectTimeout conn connect-timeout-ms)
-      (.setReadTimeout conn read-timeout-ms)
+      (.setConnectTimeout conn default-connect-timeout-ms)
+      (.setReadTimeout conn default-read-timeout-ms)
       {:status (.getResponseCode conn)
        :body (.getInputStream conn)
        :content-type (.getContentType conn)})